ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
motor>=3.7.0
pymongo>=4.0.0
redis==5.0.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
requests==2.31.0
python-multipart==0.0.6
reportlab==4.0.7
openpyxl==3.1.2
prophet==1.1.5
tensorflow-cpu==2.13.0

